+++ src/Enterspeed.Delivery.Sdk/Api/Models/DeliveryQueryBuilder.cs
@@ -13,6 +13,7 @@ namespace Enterspeed.Delivery.Sdk.Api.Models
         private readonly IList<string> _handles = new List<string>();
         private readonly IList<string> _ids = new List<string>();
         private string _url;
+        private bool _isDeliveryApiUrl;
 
         public DeliveryQueryBuilder WithHandle(string handle)
         {
@@ -47,9 +48,21 @@ namespace Enterspeed.Delivery.Sdk.Api.Models
             return this;
         }
 
+        public DeliveryQueryBuilder WithDeliveryApiUrl(string url)
+        {
+            if (_url != null && url != _url)
+            {
+                throw new EnterspeedDeliveryException("Only one URL is allowed.");
+            }
+
+            _isDeliveryApiUrl = true;
+            _url = url;
+            return this;
+        }
+
         internal DeliveryQuery Build()
         {
-            return new DeliveryQuery(_url, _handles, _ids);
+            return new DeliveryQuery(_url, _handles, _ids, _isDeliveryApiUrl);
         }
     }
 }